clean code refactoring
