+++ frontend/pages/contacts/components/tableContacts.tsx
@@ -1,31 +1,12 @@
-// svgs
-import { ReactComponent as PencilIcon } from "@assets/svg/files/pencil.svg";
-import { ReactComponent as TrashIcon } from "@assets/svg/files/trash-icon.svg";
-import { ReactComponent as ChevIcon } from "@assets/svg/files/chev-icon.svg";
-import { ReactComponent as CheckIcon } from "@assets/svg/files/edit-check.svg";
-import { ReactComponent as CloseIcon } from "@assets/svg/files/close.svg";
-//
 import { useTranslation } from "react-i18next";
-import {
-  AssetContact,
-  Contact,
-  NewContactSubAccount,
-  SubAccountContact,
-  SubAccountContactErr,
-} from "@redux/models/ContactsModels";
+import { Contact, NewContactSubAccount, SubAccountContact, SubAccountContactErr } from "@redux/models/ContactsModels";
 import { AccountIdentifier } from "@dfinity/nns";
-import ContactAssetPop from "./contactAssetPop";
-import { getContactColor, getInitialFromName, shortAddress } from "@/utils";
 import TableAssets from "./tableAssets";
-import { ChangeEvent, Fragment } from "react";
-import { CustomCopy } from "@components/CopyTooltip";
-import { CustomInput } from "@components/Input";
-import { clsx } from "clsx";
+import { Fragment } from "react";
 import { DeleteContactTypeEnum } from "@/const";
 import { Principal } from "@dfinity/principal";
-import { GeneralHook } from "@pages/home/hooks/generalHook";
 import { useContacts } from "../hooks/contactsHook";
-import { AssetToAdd } from "@redux/models/AccountModels";
+import ContactRow from "./contactRow";
 
 interface TableContactsProps {
   changeName(value: string): void;
@@ -55,15 +36,12 @@ const TableContacts = ({
   assetFilter,
 }: TableContactsProps) => {
   const { t } = useTranslation();
-
-  const { assets, getAssetIcon } = GeneralHook();
   const {
     contacts,
     selContactPrin,
     setSelContactPrin,
     checkPrincipalValid,
     updateContact,
-    addAsset,
     selCntcPrinAddAsst,
     setSelCntcPrinAddAsst,
     contactEdited,
@@ -102,114 +80,24 @@ const TableContacts = ({
       <tbody>
         {getContactsToShow().map((cntc, k) => (
           <Fragment key={k}>
-            <tr className={contactStyle(cntc)}>
-              <td className="">
-                <div className="relative flex flex-row justify-start items-center w-full min-h-14 gap-2 px-4">
-                  {(cntc.principal === selContactPrin ||
-                    cntc.principal === openAssetsPrin ||
-                    cntc.principal === selCntcPrinAddAsst) && (
-                    <div className="absolute left-0 w-1 h-14 bg-SelectRowColor"></div>
-                  )}
-                  {cntc.principal === selContactPrin ? (
-                    <CustomInput
-                      intent={"primary"}
-                      border={contactEditedErr.name ? "error" : "selected"}
-                      sizeComp={"xLarge"}
-                      sizeInput="small"
-                      value={contactEdited.name}
-                      onChange={onContactNameChange}
-                    />
-                  ) : (
-                    <div className="flex flex-row justify-start items-center w-full gap-2">
-                      <div
-                        className={`flex justify-center items-center !min-w-[2rem] w-8 h-8 rounded-md ${getContactColor(
-                          k,
-                        )}`}
-                      >
-                        <p className="text-PrimaryTextColor">{getInitialFromName(cntc.name, 2)}</p>
-                      </div>
-                      <p className="text-left opacity-70 break-words w-full max-w-[14rem]">{cntc.name}</p>
-                    </div>
-                  )}
-                </div>
-              </td>
-              <td className="py-2">
-                <div className="flex flex-row justify-start items-center gap-2 opacity-70 px-2">
-                  <p>{shortAddress(cntc.principal, 12, 9)}</p>
-                  <CustomCopy size={"xSmall"} className="p-0" copyText={cntc.principal} />
-                </div>
-              </td>
-              <td className="py-2">
-                <div className="flex flex-row justify-center items-center w-full">
-                  <div
-                    className={
-                      "flex flex-row justify-between items-center w-28 h-8 rounded bg-black/10 dark:bg-white/10"
-                    }
-                  >
-                    <p className="ml-2">{`${cntc.assets.length} ${t("assets")}`}</p>
-                    <ContactAssetPop
-                      compClass="flex flex-row justify-center items-center"
-                      btnClass="!w-8 !h-8 bg-AddSecondaryButton rounded-l-none"
-                      assets={getFilteredAssets(cntc)}
-                      getAssetIcon={getAssetIcon}
-                      onAdd={(data) => {
-                        onAddAssets(data, cntc);
-                      }}
-                      onOpen={() => {
-                        onAddAssetPopOpen(cntc);
-                      }}
-                      onClose={() => {
-                        setSelCntcPrinAddAsst("");
-                      }}
-                    />
-                  </div>
-                </div>
-              </td>
-              <td className="py-2">
-                <div className="flex flex-row justify-center items-start gap-4 w-full">
-                  {cntc.principal === selContactPrin ? (
-                    <CheckIcon
-                      onClick={() => {
-                        onSave(cntc);
-                      }}
-                      className="w-4 h-4 stroke-PrimaryTextColorLight dark:stroke-PrimaryTextColor opacity-50 cursor-pointer"
-                    />
-                  ) : (
-                    <PencilIcon
-                      onClick={() => {
-                        onEdit(cntc);
-                      }}
-                      className="w-4 h-4 fill-PrimaryTextColorLight dark:fill-PrimaryTextColor opacity-50 cursor-pointer"
-                    />
-                  )}
-                  {cntc.principal === selContactPrin ? (
-                    <CloseIcon
-                      onClick={onClose}
-                      className="w-5 h-5 stroke-PrimaryTextColorLight dark:stroke-PrimaryTextColor opacity-50 cursor-pointer"
-                    />
-                  ) : (
-                    <TrashIcon
-                      onClick={() => {
-                        onDelete(cntc);
-                      }}
-                      className="w-4 h-4 fill-PrimaryTextColorLight dark:fill-PrimaryTextColor cursor-pointer"
-                    />
-                  )}
-                </div>
-              </td>
-              <td className="py-2">
-                <div className="flex flex-row justify-center items-start gap-2 w-full">
-                  <ChevIcon
-                    onClick={() => {
-                      onChevIconClic(cntc);
-                    }}
-                    className={`w-8 h-8 stroke-PrimaryTextColorLight dark:stroke-PrimaryTextColor stroke-0  cursor-pointer ${
-                      cntc.principal === openAssetsPrin ? "" : "rotate-90"
-                    }`}
-                  />
-                </div>
-              </td>
-            </tr>
+            <ContactRow
+              cntc={cntc}
+              k={k}
+              selContactPrin={selContactPrin}
+              openAssetsPrin={openAssetsPrin}
+              selCntcPrinAddAsst={selCntcPrinAddAsst}
+              contactEditedErr={contactEditedErr}
+              contactEdited={contactEdited}
+              setContactEdited={setContactEdited}
+              setContactEditedErr={setContactEditedErr}
+              onAddAssetPopOpen={onAddAssetPopOpen}
+              setSelCntcPrinAddAsst={setSelCntcPrinAddAsst}
+              onSave={onSave}
+              onEdit={onEdit}
+              onClose={onClose}
+              onDelete={onDelete}
+              onChevIconClic={onChevIconClic}
+            />
             {cntc.principal === openAssetsPrin && (
               <tr className="bg-SecondaryColorLight dark:bg-SecondaryColor">
                 <td colSpan={5} className="w-full h-4 border-BorderColorTwoLight dark:border-BorderColorTwo">
@@ -274,45 +162,12 @@ const TableContacts = ({
       }
     });
   }
-
-  function onContactNameChange(e: ChangeEvent<HTMLInputElement>) {
-    setContactEdited((prev: any) => {
-      return { ...prev, name: e.target.value };
-    });
-    setContactEditedErr((prev: any) => {
-      return { name: false, principal: prev.principal };
-    });
-  }
-
-  function getFilteredAssets(cntc: Contact) {
-    return assets.filter((ast) => {
-      let isIncluded = false;
-      cntc.assets.map((contAst) => {
-        if (ast.tokenSymbol === contAst.tokenSymbol) isIncluded = true;
-      });
-      return !isIncluded;
-    });
-  }
-
-  function onAddAssets(data: AssetToAdd[], cntc: Contact) {
-    const auxAsst: AssetContact[] = data.map((dt) => {
-      return {
-        symbol: dt.symbol,
-        tokenSymbol: dt.tokenSymbol,
-        logo: dt.logo,
-        subaccounts: [],
-      };
-    });
-    addAsset(auxAsst, cntc.principal);
-  }
-
   function onAddAssetPopOpen(cntc: Contact) {
     setSelCntcPrinAddAsst(cntc.principal);
     setAddSub(false);
     setSelSubaccIdx("");
     setSelContactPrin("");
   }
-
   function onSave(cntc: Contact) {
     setContactEditedErr({
       name: contactEdited.name.trim() === "",
@@ -336,7 +191,6 @@ const TableContacts = ({
       setSelContactPrin("");
     }
   }
-
   function onEdit(cntc: Contact) {
     setAddSub(false);
     setSelSubaccIdx("");
@@ -347,12 +201,10 @@ const TableContacts = ({
     }
     setContactEditedErr({ name: false, principal: false });
   }
-
   function onClose() {
     setSelContactPrin("");
     setSubaccEditedErr({ name: false, subaccount_index: false });
   }
-
   function onDelete(cntc: Contact) {
     setAddSub(false);
     setSelContactPrin("");
@@ -374,7 +226,6 @@ const TableContacts = ({
     });
     setDeleteModal(true);
   }
-
   function onChevIconClic(cntc: Contact) {
     if (cntc.principal === openAssetsPrin) setOpenAssetsPrin("");
     else {
@@ -387,16 +238,6 @@ const TableContacts = ({
     setOpenSubaccToken("");
     setSelSubaccIdx("");
     setAddSub(false);
-  }
-
-  // Tailwind CSS
-  function contactStyle(cntc: Contact) {
-    return clsx({
-      ["border-b border-BorderColorTwoLight dark:border-BorderColorTwo"]: true,
-      ["bg-SelectRowColor/10"]: cntc.principal === selContactPrin || cntc.principal === selCntcPrinAddAsst,
-      ["bg-SecondaryColorLight dark:bg-SecondaryColor"]:
-        cntc.principal === openAssetsPrin && cntc.principal !== selContactPrin && cntc.principal !== selCntcPrinAddAsst,
-    });
   }
 };
