+++ frontend/pages/home/components/ICRC/SubaccountAction.tsx
@@ -53,17 +53,6 @@ const ICRCSubaccountAction = ({ setDrawerOption, setDrawerOpen }: ICRCSubaccount
           </div>
           <p className="text-md">{t("receive")}</p>
         </div>
-        {/* <div className="flex flex-col justify-center items-center w-1/3 gap-1">
-          <div
-            className="flex flex-row justify-center items-center w-7 h-7 bg-SelectRowColor rounded-md cursor-pointer"
-            onClick={() => {
-              setDrawer(DrawerOptionEnum.Enum.WRAP);
-            }}
-          >
-            <img src={WrapIcon} className="w-4 h-4" alt="wrap-icon" />
-          </div>
-          <p className="text-md">{t("wrap")}</p>
-        </div> */}
       </div>
     </Fragment>
   );
